+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ##### Enhancements
 
-* None.  
+* Add `ancestor` to return the outermost spec around a subspec  
+  [Paul Beusterien](https://github.com/paulb777)
+  [#407](https://github.com/CocoaPods/Core/pull/407)
 
 ##### Bug Fixes